Puts settings into header and moves a lot more stuff out of App.vue
